test: split close reason coverage
