+++ test/helpers.ts
@@ -1,0 +1,189 @@
+import { mkdirSync, writeFileSync } from "node:fs";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+
+export const tmpPrefix = join(tmpdir(), "clawsweeper-test-");
+
+export function item(overrides = {}) {
+  return {
+    repo: "openclaw/openclaw",
+    number: 123,
+    kind: "issue",
+    title: "Sample item",
+    url: "https://github.com/openclaw/openclaw/issues/123",
+    createdAt: "2026-01-01T00:00:00Z",
+    updatedAt: "2026-01-01T00:00:00Z",
+    author: "contributor",
+    authorAssociation: "NONE",
+    labels: [],
+    ...overrides,
+  };
+}
+
+export function closeDecision(overrides = {}) {
+  return {
+    decision: "close",
+    closeReason: "implemented_on_main",
+    confidence: "high",
+    summary: "Current main already implements this.",
+    changeSummary: "Requests confirmation that the feature works on current main.",
+    evidence: [
+      {
+        label: "implementation",
+        detail: "The feature is present in source.",
+        file: "src/example.ts",
+        line: 12,
+        command: null,
+        sha: "abcdef1234567890",
+      },
+      {
+        label: "git history provenance",
+        detail: "git blame traces the implemented line to abcdef1234567890.",
+        file: "src/example.ts",
+        line: 12,
+        command: "git blame -L 12,12 -- src/example.ts",
+        sha: "abcdef1234567890",
+      },
+      {
+        label: "release provenance",
+        detail: "The fix is on current main and no containing release tag was found.",
+        file: null,
+        line: null,
+        command: "git tag --contains abcdef1234567890",
+        sha: "abcdef1234567890",
+      },
+    ],
+    likelyOwners: [
+      {
+        person: "@alice",
+        role: "introduced behavior",
+        reason: "git blame points the relevant implementation line at abcdef1234567890.",
+        commits: ["abcdef1234567890"],
+        files: ["src/example.ts"],
+        confidence: "high",
+      },
+      {
+        person: "@bob",
+        role: "recent maintainer",
+        reason: "Recent adjacent commits changed the same code path.",
+        commits: ["1234567890abcdef"],
+        files: ["src/example.ts"],
+        confidence: "medium",
+      },
+    ],
+    risks: [],
+    bestSolution: "Keep the implementation as-is.",
+    triagePriority: "P2",
+    impactLabels: [],
+    mergeRiskLabels: [],
+    mergeRiskOptions: [],
+    reviewMetrics: [],
+    labelJustifications: [
+      {
+        label: "P2",
+        reason: "Normal priority applies to this limited-scope implemented behavior check.",
+      },
+    ],
+    itemCategory: "bug",
+    reproductionStatus: "reproduced",
+    reproductionConfidence: "high",
+    requiresNewFeature: false,
+    requiresNewConfigOption: false,
+    requiresProductDecision: false,
+    reproductionAssessment:
+      "Yes. Current main can be checked by inspecting src/example.ts and git blame evidence.",
+    solutionAssessment:
+      "Yes. Keeping the implementation as-is is the narrowest maintainable outcome.",
+    visionFit: "not_applicable",
+    visionFitReason: "Vision-fit assessment is not needed for this implemented close decision.",
+    visionFitEvidence: [],
+    implementationComplexity: "not_applicable",
+    autoImplementationCandidate: "none",
+    rootCauseCluster: {
+      confidence: "low",
+      canonicalRef: null,
+      currentItemRelationship: "independent",
+      summary: "No evidence-backed root-cause cluster was established.",
+      members: [],
+    },
+    agentsPolicyStatus: {
+      found: true,
+      readFully: true,
+      applied: true,
+      status: "found_applied",
+      summary: "Found AGENTS.md and applied relevant repository review guidance.",
+    },
+    reviewFindings: [],
+    securityReview: {
+      status: "not_applicable",
+      summary: "No patch security review is needed for this issue cleanup decision.",
+      concerns: [],
+    },
+    realBehaviorProof: {
+      status: "not_applicable",
+      summary: "Real behavior proof is not required for non-PR issue triage.",
+      evidenceKind: "not_applicable",
+      needsContributorAction: false,
+    },
+    prRating: {
+      proofTier: "NA",
+      patchTier: "NA",
+      overallTier: "NA",
+      summary: "PR readiness rating is not applicable to this issue cleanup decision.",
+      nextSteps: [],
+    },
+    telegramVisibleProof: {
+      status: "not_needed",
+      summary: "This non-PR issue triage does not need Telegram visible proof.",
+    },
+    mantisRecommendation: {
+      status: "not_recommended",
+      scenario: "none",
+      reason: "Mantis proof is not useful for this issue triage.",
+      maintainerComment: "",
+    },
+    featureShowcase: {
+      status: "none",
+      reason: "This item is not an unusually compelling feature idea.",
+    },
+    overallCorrectness: "not a patch",
+    overallConfidenceScore: 0.75,
+    fixedRelease: null,
+    fixedSha: "abcdef1234567890",
+    fixedAt: "2026-04-28T12:00:00Z",
+    closeComment: "Closing this as implemented after Codex review.\n\n- Evidence.",
+    workCandidate: "none",
+    workConfidence: "low",
+    workPriority: "low",
+    workReason: "Close decisions do not need a fix PR.",
+    workPrompt: "",
+    workClusterRefs: [],
+    workValidation: [],
+    workLikelyFiles: [],
+    ...overrides,
+  };
+}
+
+export const git = {
+  mainSha: "abcdef1234567890",
+  latestRelease: null,
+};
+
+export function withMockGh(root: string, script: string, run: () => void): void {
+  const originalGhBin = process.env.GH_BIN;
+  const originalGhBinArgs = process.env.GH_BIN_ARGS;
+  const binDir = join(root, "bin");
+  mkdirSync(binDir, { recursive: true });
+  const ghPath = join(binDir, "gh.js");
+  writeFileSync(ghPath, script, { mode: 0o755 });
+  try {
+    process.env.GH_BIN = process.execPath;
+    process.env.GH_BIN_ARGS = JSON.stringify([ghPath]);
+    run();
+  } finally {
+    if (originalGhBin === undefined) delete process.env.GH_BIN;
+    else process.env.GH_BIN = originalGhBin;
+    if (originalGhBinArgs === undefined) delete process.env.GH_BIN_ARGS;
+    else process.env.GH_BIN_ARGS = originalGhBinArgs;
+  }
+}
